Add child component for calculator and make changes on style
