+++ src/App.css
@@ -10,16 +10,15 @@ body {
   background-color: burlywood;
   display: grid;
   place-content: center;
-  font-size: 2rem;
+  font-size: 1.3rem;
   height: 100vh;
 }
 
 .container {
-  height: 75vh;
-  width: 50vh;
-  background-color: black;
+  height: 65vh;
+  width: 60vh;
+  background-color: rgba(110, 111, 111, 0.977);
   border-radius: 1rem;
-  padding: 1rem;
   -webkit-box-shadow: 7px 7px 5px 0 rgba(50, 50, 50, 0.75);
   -moz-box-shadow: 7px 7px 5px 0 rgba(50, 50, 50, 0.75);
   box-shadow: 7px 7px 5px 0 rgba(50, 50, 50, 0.75);
@@ -31,28 +30,28 @@ body {
   width: 100%;
   grid-template-columns: repeat(4, 1fr);
   grid-template-rows: repeat(6, 1fr);
-  gap: 10px;
 }
 
 .screen {
   grid-column: 1/-1;
   color: white;
-  text-align: center;
-  padding: 0.5rem;
-  font-size: 4rem;
+  text-align: end;
+  padding-right: 10px;
+  padding-top: 18px;
+  font-size: 2rem;
   overflow: hidden;
-  background-color: rgba(43, 40, 40, 0.993);
-  border-radius: 4px;
+  background-color: rgba(134, 132, 132, 0.703);
 }
 
 .btn {
   display: grid;
   place-content: center;
-  background-color: rgb(54, 52, 52);
-  border-radius: 50%;
-  color: white;
+  background-color: rgb(220, 220, 220);
+  border-radius: 1px;
+  color: black;
   cursor: pointer;
   font-weight: bold;
+  border: 1px solid rgb(185, 185, 185);
 }
 
 .btn:hover {
@@ -60,7 +59,7 @@ body {
 }
 
 .light-gray {
-  background-color: gray;
+  background-color: rgb(220, 220, 220);
   color: black;
 }
 
@@ -70,7 +69,6 @@ body {
 
 .orange {
   background-color: orange;
-  color: white;
 }
 
 .orange:hover {
@@ -80,7 +78,6 @@ body {
 
 .zero {
   grid-column: 1/span 2;
-  border-radius: 25px;
   place-content: unset;
   align-items: center;
   padding-left: 2rem;
